Found the game adding bug and created a decent workaround
There's still a bug in here because it has to do with the data structure as it goes in and out of YAML, the non-orderedness and the week getting inserted into the middle.  But the workaround works and is larger than anyone would need so probably okay for my purposes.